% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_metrics.R
\name{calculate_cluster_metrics}
\alias{calculate_cluster_metrics}
\title{Calculate Metrics for Clusters}
\usage{
calculate_cluster_metrics(cluster, verbose, page_rank_calculation, ...)
}
\arguments{
\item{cluster}{Cluster as provided within the function}

\item{verbose}{Print out Metrics?}

\item{...}{Additional parameters of the clustering algorithm}
}
\value{
Metrics
}
\description{
Internal function that calculates the metrics of clusters calculated during the topic calculations
}
\keyword{internal}
