% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_network.R
\name{calculate_network}
\alias{calculate_network}
\title{Calculate PMI Weighted Network}
\usage{
calculate_network(
  data,
  document,
  feature,
  pmi_weight = TRUE,
  as_data_frame = FALSE,
  keep_negative_weights = TRUE
)
}
\arguments{
\item{data}{Data frame containing the network data.}

\item{document}{Name of the vertex A column.}

\item{feature}{Name of the vertex B column.}

\item{pmi_weight}{Logical indicating whether weights should be calculated based on the PMI (Pointwise Mutual Information) of `document` and `feature`. If `FALSE`, a simple cooccurrence weighting is performed.}

\item{as_data_frame}{Logical indicating whether to output a data frame with merged edges (a to b | b to a)}

\item{keep_negative_weights}{Logical indicating whether to keep edges with negative PMI weights. Only applies if `pmi_weight = TRUE`.}
}
\value{
A weighted igraph object if pmi_weight is TRUE, otherwise an igraph object or data frame depending on as_data_frame.
}
\description{
This function calculates a pointwise mutual information (PMI) weighted network from vertex data.
}
\keyword{internal}
