% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_topics.R
\name{calculate_topics}
\alias{calculate_topics}
\title{Calculate Topic Cluster from a Text Network}
\usage{
calculate_topics(
  text_network,
  documents = NULL,
  document_tokens = NULL,
  document_ids = NULL,
  negative_edge_weights = TRUE,
  page_rank_calculation = c("global", "cluster"),
  cluster_function = igraph::cluster_leiden,
  ...,
  keep_cluster_object = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{text_network}{An igraph graph representing the text network.}

\item{documents}{Optional, a data frame containing document information. Expects one row per token. NULL to skip.
Returns additional information on topic-relevant documents if provided.}

\item{document_tokens}{String; The column name in documents containing the document tokens.}

\item{document_ids}{String; The column name in documents containing document identifiers.}

\item{negative_edge_weights}{Logical indicating whether to consider only edges
with positive weights. If the chosen clustering algorithm does not support negative edge weights,
set this to `FALSE`.}

\item{page_rank_calculation}{Character specifying the type of Page Rank calculation.
Options are "global" for global network Page Rank, or 
"cluster" for calculating the Page Rank of the entities within each topic cluster.}

\item{cluster_function}{The clustering function to use. Can be any igraph clustering function.
Default is `igraph::cluster_leiden`.}

\item{...}{Additional arguments passed to `cluster_function`.}

\item{keep_cluster_object}{Logical indicating whether to keep the igraph cluster object.
Can be helpful for additional checks. See `help(igraph::membership)`}

\item{verbose}{Logical indicating whether to print clustering metrics.}
}
\value{
A list object of class "textgraph_topics" containing topic entities, 
  document data (optional), clustering metrics, and the cluster object (optional).
}
\description{
This function takes a text network represented as an igraph graph and performs
topic clustering using the specified clustering algorithm. It then calculates
additional metrics such as Page Rank, cluster metrics, and more.
}
\details{
The function performs topic clustering on the input text network, calculates
additional metrics based on the specified parameters, and returns a structured object
containing topics and metrics.
}
\examples{
\dontrun{
topics <- calculate_topics(text_network,
                          documents = NULL,
                          document_tokens = NULL,
                          document_ids = NULL,
                          negative_edge_weights = TRUE,
                          page_rank_calculation = "global",
                          cluster_function = cluster_leiden,
                          objective_function = "CPM",
                          keep_cluster_object = FALSE,
                          verbose = TRUE)
# optionally, we can add a dataframe with document information to get relevant documents
data("de_pol_twitter")
topics <- calculate_topics(text_network,
                          documents = de_pol_twitter,
                          document_tokens = "lemma",
                          document_ids = "doc_id",
                          negative_edge_weights = TRUE,
                          page_rank_calculation = "global",
                          cluster_function = cluster_leiden,
                          objective_function = "CPM",
                          keep_cluster_object = FALSE,
                          verbose = TRUE)
}

}
