% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.adjacency.matrix.mono.R
\name{compute.adjacency.matrix.mono}
\alias{compute.adjacency.matrix.mono}
\title{Compute Adjacency Matrix for a Monoplex Network}
\usage{
compute.adjacency.matrix.mono(x)
}
\arguments{
\item{x}{A multiplex network object.}
}
\value{
A dgCMatrix object representing the adjacency matrix of the network.
}
\description{
This helper function computes the adjacency matrix for a monoplex network.
}
\details{
An adjusted version of the compute.adjacency.matrix() function from
 the RandomWalkRestartMH package, making monoplex network preparation
 more efficient by dropping unnecessary overhead
 Specifically, it vastly reduces the use of RAM (which would require
 manual flushing with gc() every time) by dropping everything connected
 to the line "offdiag <- (delta/(L-1))*Idem_Matrix" which is not needed
 for monoplex networks
}
\keyword{internal}
