% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_quantile.R
\name{drop_quantile}
\alias{drop_quantile}
\title{Drop lower quantiles of tokens}
\usage{
drop_quantile(
  data,
  tokens,
  quantile = 0.1,
  ignore_case = TRUE,
  group,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data Frame}

\item{tokens}{String; column containing the tokens to count and drop}

\item{quantile}{Numerical; the quantile up to and including which tokens are dropped}

\item{ignore_case}{Logical; should upper/lower case of the tokens be ignored? Preserves original case}

\item{group}{String; grouping column, e.g. type of word, for report statistics}

\item{verbose}{Logical; should statistics be printed out?}
}
\value{
A data frame containing all the variables of the original `data` as well as the additional 
 variable `entity count` indicating the number of times a token appears in the data. All tokens below 
 the `entity_count` specified through the quantile (and printed through `verbose`) are dropped.
}
\description{
This function drops all tokens in a data frame occurring below the threshold of a given quantile
}
\details{
Tokens falling within or below the specified quantile will be dropped. E.g., when specifying
 `quantile = 0.1`, all tokens within or below the 10% quantile of the entity count will be dropped.
}
\examples{
data("de_pol_twitter")

de_pol_twitter_reduced <- drop_quantile(data = de_pol_twitter,
                                       tokens = "lemma",
                                       quantile = 0.1,
                                       ignore_case = FALSE, # in this case, "lemma" has already been set to lower case
                                       group = "tag",
                                       verbose = TRUE)
}
