% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_topics.R
\name{filter_topics}
\alias{filter_topics}
\title{Filter the topics in a textgraph_topics object}
\usage{
filter_topics(textgraph_topics, topic_numbers)
}
\arguments{
\item{textgraph_topics}{A textgraph_topics object, as provided by the `calculate_topics()` function.}

\item{topic_numbers}{A numerical vector containing the numbers of the topics to be retained.}
}
\value{
A textgraph_topics object with updated contents and metrics.
}
\description{
This function allows to filter the topics in a textrgraph topic by providing a
 numerical vector of topics to be retained. Note that this also updates the metrics
 where applicable (see Details).
}
\details{
The returned textgraph_topics object has its metrics updated where applicable.
 Specifically, metrics related to the clustering algorithm are NOT updated, but conserved
 from the original topic calculation. Metrics concerning the number of topics, and their mean
 and median number of entities, however, are updated.
}
\examples{
\dontrun{
filtered_topics <- filter_topics(textgraph_topics = topics,
                                 topic_numbers = c(1, 14, 76))
}

}
