% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Calculate Geometric Mean}
\usage{
geometric.mean(Scores, L, N)
}
\arguments{
\item{Scores}{Scores}

\item{L}{L}

\item{N}{N}
}
\value{
Score
}
\description{
Calculate Geometric Mean
}
\details{
An internal function from the RandomWalkRestartMH package, required for
  the RWR calculations. Pulled from https://github.com/alberto-valdeolivas/RandomWalkRestartMH/blob/master/R/InternalFunctions.R
}
\keyword{internal}
