% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seed.scoresMultiplex.R
\name{get.seed.scoresMultiplex}
\alias{get.seed.scoresMultiplex}
\title{Get Scores for the seeds of the multiplex network}
\usage{
get.seed.scoresMultiplex(Seeds, Number_Layers, tau)
}
\arguments{
\item{Seeds}{Seeds}

\item{Number_Layers}{Layers}

\item{tau}{tau}
}
\value{
Scores
}
\description{
Get Scores for the seeds of the multiplex network
}
\details{
An internal function from the RandomWalkRestartMH package, required for
  the RWR calculations. Pulled from https://github.com/alberto-valdeolivas/RandomWalkRestartMH/blob/master/R/InternalFunctions.R
}
\keyword{internal}
