% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unclassified_documents.R
\name{get_unclassified_documents}
\alias{get_unclassified_documents}
\title{Get Unclassified Documents}
\usage{
get_unclassified_documents(
  classification_result,
  documents,
  doc_id,
  n = 20,
  mode = c("print", "return")
)
}
\arguments{
\item{classification_result}{The result of `classify_documents()`. Must contain
unclassified documents (`return_unclassified_docs = TRUE` in `classify_document()`).}

\item{documents}{The full document data to be matched to the classification result for output. Usually the non-tokenized documents.}

\item{doc_id}{The name of the document identifier used for matching.
Can be a `dplyr::join_by()` function where `classification_result` ID = `a` and `documents` ID = `b`.}

\item{n}{The number of documents to print or return. If larger or equal to the number of unclassified documents,
all documents will be returned, rather than a random sample (default: 20).}

\item{mode}{The mode to operate in, either "print" to print results or
"return" to return a data frame of results (default: "print").}
}
\value{
Depending on `mode`, either prints the results or
  returns a data frame containing the specified number of missing documents.
}
\description{
A function to return or print a random sample of documents not classified by `classify_documents()`.
}
\examples{
# the classified_documents object in this example is created by classify_document() - see help(classify_documents)
# note that you would usually want to use non-tokenized documents as 'documents'. This data, however, does not come
#   with the package. Instead, we paste the tokens of each document together

data("de_pol_twitter")

get_unclassified_documents(classification_result = classified_documents,
                           documents = de_pol_twitter \%>\% dplyr::summarise(content = paste(lemma, collapse = ","), .by = doc_id),
                           doc_id = "doc_id",
                           n = 20,
                           mode = "print")

}
