% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_community_matching.R
\name{install_community_matching}
\alias{install_community_matching}
\title{Install the Community Matching Algorithm required for Dynamic Topic Matching}
\usage{
install_community_matching(
  envname = "textgraph",
  method = "auto",
  install_dependencies = TRUE,
  ...
)
}
\arguments{
\item{envname}{String; Name of the environment used for installation. Defaults to a specific "textgraph" environment}

\item{method}{See [reticulate::py_install()]}

\item{install_dependencies}{Logical; Should additionally required packages be installed as well? Note that
the required three packages `numpy`, `pandas`, and `scipy` are set to fixed versions to avoid future conflicts.
Recommended to set this to `false` if using an existing environment for `envname` to not overwrite existing
package versions (but be prepared for errors caused by mismatching package versions).}

\item{...}{Additional arguments to pass on to [reticulate::py_install()]}
}
\value{
Sets up a reticulate environment
}
\description{
Installs the required packages and the matching algorithm into a specified reticulate Python environment.
Generally recommended to use the fresh `textgraph` environment when setting up.
}
\examples{
\dontrun{
install_community_matching()
}
}
