% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_textgraph_topics.R
\name{load_textgraph_topics}
\alias{load_textgraph_topics}
\title{Load textgraph topics}
\usage{
load_textgraph_topics(file, verbose)
}
\arguments{
\item{file}{The `.tar.gz` file to load}
}
\value{
Returns the `textgraph_topics` object
}
\description{
A function to load the textgraph_topics object created by `calculate_topics()`
  or `calculate_dynamic_topics()` and saved via `save_textgraph_topics()`
}
\examples{
\dontrun{
my_textgraph_topics <- load_textgraph_topics("my_textgraph_topics.tar.gz")
}

}
