% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_multiplex_objects.R
\name{make_multiplex_objects}
\alias{make_multiplex_objects}
\title{Create Multiplex Network Objects}
\usage{
make_multiplex_objects(
  dat,
  document,
  feature,
  pmi_weight = TRUE,
  keep_negative_weights = TRUE,
  network = NULL,
  keep_igraph_network = FALSE,
  keep_multiplex_network = TRUE,
  keep_adjacency_matrix = FALSE,
  keep_normalized_adjacency_matrix = TRUE
)
}
\arguments{
\item{dat}{Data frame to pass to calculate_network function. Expects a dataframe in the format of an edge list, where each row depicts an edge between `document` and `feature`.}

\item{document}{String; Name of the document ID column.}

\item{feature}{String; Name of the feature (e.g. token) column.}

\item{pmi_weight}{Logical indicating whether weights should be calculated based on the PMI (Pointwise Mutual Information) of `document` and `feature`. If `FALSE`, a simple cooccurrence weighting is performed.}

\item{keep_negative_weights}{Logical indicating whether edges with negative PMI weight should be retained. Discarding them can make for a smaller network.}

\item{network}{Pre-calculated network to turn into a multiplex network; if `NULL`, a network will be calculated from the vertices. Expects either `NULL` or an igraph graph.}

\item{keep_igraph_network}{Logical indicating whether the igraph network should be kept separately.}

\item{keep_multiplex_network}{Logical indicating whether the multiplex network should be kept separately.}

\item{keep_adjacency_matrix}{Logical indicating whether the non-normalized adjacency matrix should be kept.}

\item{keep_normalized_adjacency_matrix}{Logical indicating whether to keep the normalized adjacency matrix.}
}
\value{
A list containing various elements such as the igraph network, multiplex network, and adjacency matrices, to be used in `get_rwr_terms()`.
}
\description{
This function creates and preprocesses a multiplex network for later use in Random Walks with Restart.
}
\details{
This function either turns a data frame of edges or an existing igraph object into a multiplex network object. The functionality to weight the network via the Pointwise
 Mutual Information (PMI) between two nodes is especially useful for text networks. Additional options to return only specific parts of the multiplex object are provided.
 Note that, for use with the `get_rwr_terms()` function, the normalized adjacency matric (`keep_normalized_adjacency_matrix = TRUE`) and the multiplex network
 (`keep_multiplex_network = TRUE`) are required. If only the igraph network is required, one can only set `keep_igraph_network` to `TRUE`. The calculation of the adjecency
 matrix will be skipped in this case.
}
\examples{
data("de_pol_twitter")

multiplex_text_network <- make_multiplex_objects(de_pol_twitter,
                                                 document = "doc_id",
                                                 feature = "lemma",
                                                 pmi_weight = TRUE,
                                                 keep_negative_weights = TRUE,
                                                 network = NULL,
                                                 keep_igraph_network = FALSE,
                                                 keep_multiplex_network = TRUE,
                                                 keep_adjacency_matrix = FALSE,
                                                 keep_normalized_adjacency_matrix = TRUE)


}
