% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_textnetwork.R
\name{make_textnetwork}
\alias{make_textnetwork}
\title{Calculate a Text Cooccurrence Network}
\usage{
make_textnetwork(
  data,
  document,
  feature,
  pmi_weight = TRUE,
  as_data_frame = FALSE,
  keep_negative_weights = TRUE
)
}
\arguments{
\item{data}{Data frame containing the network data.}

\item{document}{String; Name of the document ID column.}

\item{feature}{String; Name of the feature (e.g. token) column.}

\item{pmi_weight}{Logical indicating whether weights should be calculated based on the PMI (Pointwise Mutual Information) of `document` and `feature`. If `FALSE`, a simple cooccurrence weighting is performed.}

\item{as_data_frame}{Logical indicating whether to output a data frame with merged edges (a to b | b to a)}

\item{keep_negative_weights}{Logical indicating whether to keep edges with negative PMI weights. Only applies if `pmi_weight = TRUE`.}
}
\value{
A weighted igraph object if pmi_weight is TRUE, otherwise an igraph object or data frame depending on as_data_frame.
}
\description{
This function calculates a pointwise mutual information (PMI) weighted network from vertex data.
}
\details{
This is effectively a wrapper for the internal function `calculate_networks()`, allowing to calculate a text cooccurrence network
 from a data frame.
}
\examples{
data("de_pol_twitter")
textnetwork <- make_textnetwork(data = de_pol_twitter,
                                document = "doc_id", 
                                feature = "lemma",
                                pmi_weight = TRUE,
                                as_data_frame = FALSE,
                                keep_negative_weights = TRUE)

}
