% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random.Walk.Restart.Multiplex.failsafe.R
\name{Random.Walk.Restart.Multiplex.failsafe}
\alias{Random.Walk.Restart.Multiplex.failsafe}
\title{Random Walk with Restart Failsafe for Multiplex Networks}
\usage{
Random.Walk.Restart.Multiplex.failsafe(
  x,
  MultiplexObject,
  Seeds,
  r = 0.7,
  tau,
  MeanType = "Geometric",
  DispResults = "TopScores",
  ...
)
}
\arguments{
\item{x}{A dgCMatrix object of the adjacency matrix.}

\item{MultiplexObject}{A Multiplex object.}

\item{Seeds}{A character vector of seed nodes.}

\item{r}{The restart parameter, should be between 0 and 1.}

\item{tau}{A numerical vector specifying layer weights, summing up to 1 when divided by number of layers.}

\item{MeanType}{A string specifying the mean type to compute global scores, one of "Geometric", "Arithmetic", "Sum".}

\item{DispResults}{A string to decide results display method, either "TopScores" or "Alphabetic".}

\item{...}{Additional parameters.}
}
\value{
A list with Random Walk with Restart results and meta-information.
}
\description{
This adjusted version of the RandomWalkRestartMH function includes a failsafe for infinity values during computation.
}
\details{
An adjusted version of https://github.com/alberto-valdeolivas/RandomWalkRestartMH/blob/master/R/RWRandMatrices.R
 including a failsafe for values in the proximity vectors reaching (-)Inf
}
\keyword{internal}
