% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_textgraph_topics.R
\name{save_textgraph_topics}
\alias{save_textgraph_topics}
\title{Save textgraph topics}
\usage{
save_textgraph_topics(
  textgraph_topics,
  file,
  include_document_entities = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{textgraph_topics}{The textgraph_topics object to save}

\item{file}{String; the `.tar.gz` file to save to}

\item{include_document_entities}{Logical; Should the entities attached to the document data
get saved? Saving them can be helpful for future analyses;
excluding them can speed up the save/load process and save
disk space}

\item{verbose}{Logical; sets Verbosity}
}
\value{
writes the object under the specified `file` path. If no path is
         specified, writes the file to the current working directory
}
\description{
A function to save the textgraph_topics object created by `calculate_topics()`
  or `calculate_dynamic_topics()`. Will save as `.tar.gz` file
}
\examples{
\dontrun{
save_textgraph_topics(my_textgraph_topics, "my_textgraph_topics.tar.gz")
}

}
