% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_walk.R
\name{seed_walk}
\alias{seed_walk}
\title{Run Random Walk with Restart for Each Seed Node}
\usage{
seed_walk(seed, walk_network, normalize_score, positive_scores_only)
}
\arguments{
\item{seed}{A single seed node identifier.}

\item{walk_network}{An object containing the multiplex network and various matrices.}

\item{normalize_score}{How scores should be normalized.}

\item{positive_scores_only}{Logical indicating whether to drop negative scores before normalization.}
}
\value{
A dataframe with Random Walk with Restart results for the given seed node.
}
\description{
This helper function runs Random Walk with Restart for each seed node in a given network. Used internally by `get_rwr_terms()`.
}
\keyword{internal}
