% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_group_terms.R
\name{top_group_terms}
\alias{top_group_terms}
\title{Print or Return Top Terms per Group}
\usage{
top_group_terms(
  classification_result,
  group_name,
  classification_measure,
  include_seed_terms = TRUE,
  n = 20,
  with_ties = TRUE,
  mode = c("print", "return")
)
}
\arguments{
\item{classification_result}{The result of `classify_documents()`. Must contain
walk terms (`return_walk_terms = TRUE` in `classify_document()`).}

\item{group_name}{String; The name of the group variable.}

\item{classification_measure}{String; The classification measure used for selecting
the top terms. Must be present in the data. Provided by `get_rwr_terms()` and selected as
`classification_measure` in `classify_documents()`.}

\item{include_seed_terms}{Logical indicating whether to include seed terms in the output
(default: TRUE).}

\item{n}{The maximum number of terms to return per group (default: 20).}

\item{with_ties}{Logical indicating tie handling. If TRUE, will include
tied values, potentially leading to more results returned than specified with `n`.
If FALSE, will select tied values at random (default: TRUE).}

\item{mode}{The mode to operate in, either "print" to print results or
"return" to return a data frame of results (default: "print").}
}
\value{
Depending on `mode`, either prints the results or
  returns a data frame containing the specified number of top terms for each group.
}
\description{
This function displays or returns the top terms for each group based on
 the results of document classification.
}
\examples{
\dontrun{
# the classified_documents object in this example is created by classify_document() - see help(classify_documents)

top_group_terms(classification_result = classified_documents,
                group_name = "ministry_name",
                classification_measure =  "ScoreNormMean",
                include_seed_terms = TRUE,
                n = 20,
                with_ties = TRUE,
                mode = "print")
}

}
