% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_rwr.R
\name{visualize_rwr}
\alias{visualize_rwr}
\title{Visualize Random Walk Terms}
\usage{
visualize_rwr(
  rwr_terms,
  network,
  group_name,
  classification_measure = c("Score", "ScoreMean", "ScoreNorm", "ScoreNormMean",
    "ScoreNormGroup", "ScoreNormGroupMean"),
  n_terms = 5,
  selection = c("all", "overview", c()),
  layout = "fr",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{rwr_terms}{A random walks result generated by `get_rwr_terms()`}

\item{network}{A text cooccurrence network, usually generated by
`make_multiplex_objects()`}

\item{group_name}{String; Name of the grouping variable.}

\item{classification_measure}{String; name of the measure variable to use for
classification, generated by `get_rwr_terms()`. Must be present in
`walk_terms`. One of "Score", "ScoreMean", "ScoreNorm", "ScoreNormMean",
"ScoreNormGroup", or "ScoreNormGroupMean". See `get_rwr_terms()` for
details.}

\item{n_terms}{Numerical; number of terms per group to visualize. Picks terms
with the highes `classification_measure` in descending order.}

\item{selection}{Either "all" to generate an overview plot and plots for all
groups; "overview" to only generate an overview plot; or a character vector
indicating groups to generate an overview plot and group-specific plots
for. Defaults to "all".}

\item{layout}{The layouting algorithm to use. Defaults to
Fruchterman-Reingold. See `help(ggraph::ggraph())` for options}

\item{...}{Additional layouting argument passed on to the algorithm. See
`help(ggraph::ggraph())`}

\item{verbose}{If TRUE, prints out the plots}
}
\value{
Returns a list of ggraph plots
}
\description{
Visualize Random Walk Terms
}
\examples{
\dontrun{


plots <- visualize_rwr(rwr_terms = rwr_terms,
                       network = multiplex_text_network,
                       group_name = "ministry_name",
                       classification_measure = "ScoreNormMean",
                       selection = "all",
                       n_terms = 5,
                       layout = "fr",
                       verbose = F)
}
}
