% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_document_data.R
\name{prepare_document_data}
\alias{prepare_document_data}
\title{Prepare Document Data during Topic Calculations}
\usage{
prepare_document_data(topics, documents, document_tokens, document_ids)
}
\arguments{
\item{topics}{Topics object as provided within the function workflow}

\item{documents}{document data}

\item{document_tokens}{tokens in document data}

\item{document_ids}{IDs in document data}
}
\value{
prepared document data
}
\description{
Internal function to process documents during topic calculations
}
\keyword{internal}
