% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_textgraph_topics.R
\name{save_textgraph_topics}
\alias{save_textgraph_topics}
\title{Save textgraph topics}
\usage{
save_textgraph_topics(textgraph_topics, file, verbose = TRUE)
}
\arguments{
\item{textgraph_topics}{The topics object to save}

\item{file}{The `.tar.gz` file to save to}

\item{verbose}{Verbosity}
}
\value{
writes the object under the specified `file` path. If no path is
         specified, writes the file to the current working directory
}
\description{
A function to save the textgraph_topics object created by `calculate_topics()`
  or `calculate_dynamic_topics()`. Will save as `.tar.gz` file
}
\examples{
\dontrun{
save_textgraph_topics(my_textgraph_topics, "my_textgraph_topics.tar.gz")
}

}
